Fixe typos in cfpipeline.

+++ pipeline/cfpipeline.py
@@ -2,7 +2,7 @@ import os
 import sys
 import time
 import datetime
-
+import contextlib
 from pipeline import run, mount_basespace, mount_instance_storage, unmount_basespace, list_basespace_fastqs, ungzip_and_combine_illumina_fastqs, load_panel_from_s3, \
                     s3_put, dedup, illumina_readgroup, pipe, create_report, progress
 from covermi import Panel, covermimain
@@ -22,9 +22,6 @@ def cfpipeline(basespace_project_regex="", sample_regex="", s3_project=None, pan
     threads = "4"
     
     os.chdir(mount_instance_storage())
-    if not os.path.exists("work"):
-        os.mkdir("work")
-    os.chdir("work")
     
     mount_basespace()
     fastqs = list_basespace_fastqs(project=basespace_project_regex, sample=sample_regex)
@@ -34,18 +31,18 @@ def cfpipeline(basespace_project_regex="", sample_regex="", s3_project=None, pan
 
     panel = load_panel_from_s3(panelname)
     prop = panel.properties
-    
+
     for r1_fastq, r2_fastq in zip(fastqs[::2], fastqs[1::2]):
         
-        sample = os.apth.splitext(os.path.basename(r1_fastq))
-        with open("{}_pipeline.txt".format(sample)) as f_report:
+        sample = os.path.splitext(os.path.basename(r1_fastq))[0]
+        with open("{}_pipeline.txt".format(sample), "wt") as f_report:
             with contextlib.redirect_stderr(f_report):
                 
                 start_time = time.time()
                 progress("cfPipeline {}, {}.".format(os.path.basename(r1_fastq), os.path.basename(r2_fastq)))
                 progress("Starting {}.".format(datetime.datetime.now()))
-                progress("Shaw allowed = 3, thruplex = {}, min_family_size = {}.".format(bool(prop.get("thruplex", False))), min_family_size=prop.get("min_family_size", 1))
-                dedup(r1_fastq, r2_fastq, allowed=3, thruplex=prop.get("thruplex", False), min_family_size=prop.get("min_family_size", 1))
+                progress("Shaw allowed = 3, thruplex = {}, min_family_size = {}.".format(prop.get("thruplex", False), prop.get("min_family_size", 1)))
+                dedup(r1_fastq, r2_fastq, allowed=3, thruplex=prop.get("thruplex", False), min_family_size=int(prop.get("min_family_size", 1)))
                 r1_dedupfastq = "{}.deduped.fastq".format(r1_fastq[:-6])
                 r2_dedupfastq = "{}.deduped.fastq".format(r2_fastq[:-6])
                 pipe(["wc", r1_fastq, r2_fastq, r1_dedupfastq, r2_dedupfastq], stdout=sys.stderr)
@@ -55,7 +52,7 @@ def cfpipeline(basespace_project_regex="", sample_regex="", s3_project=None, pan
                 progress("BWA mem.".format(sample))
                 sam_file = "{}.sam".format(sample)
                 with open(sam_file, "wb") as f:
-                    pipe(["bwa", "mem", "-t", threads, "-R", illumina_readgroup(r1_fastq), prop["reference_fasta"], r1_dedupfastq, r2_dedupfastq], stdout=f)
+                    pipe(["bwa", "mem", "-t", threads, "-R", illumina_readgroup(r1_dedupfastq), prop["reference_fasta"], r1_dedupfastq, r2_dedupfastq], stdout=f)
                 os.unlink(r1_dedupfastq)
                 os.unlink(r2_dedupfastq)
 
@@ -109,20 +106,22 @@ def cfpipeline(basespace_project_regex="", sample_regex="", s3_project=None, pan
                 progress("Time taken = {} seconds.".format(int(time.time() - start_time)))
 
 
+        print("Uploading to s3.")
         for filename in os.listdir():
-            s3_put(filename, prefix="{}/{}".format(s3_project, sample))
-            os.unlink(filename)
+            if os.path.isfile(filename):
+                s3_put(filename, prefix="{}/{}".format(s3_project, sample))
+                os.unlink(filename)
 
 
 if __name__ == "__main__":
-    cfpipeline(basespace_project="CAPP", sample="10010014-H3731-c-0", s3_project="Accept", panelname=panel)
+    cfpipeline(basespace_project_regex="CAPP", sample_regex="10010014-H3731-c-0", s3_project="Accept", panelname="Accept")
     sys.exit()
 
     for s3_project, panel, sample, basespace_project in [("head_and_neck", "Head_and_Neck", "HNC006-HNC006-c-0", "HNC"),
                                                          ("head_and_neck", "Head_and_Neck", "HNC011-HNC011-c-0", "HNC"),
                                                          ("head_and_neck", "Head_and_Neck", "HNC016-HNC016-c-0", "HNC"),
                                                          ]:
-        cfpipeline(basespace_project=basespace_project, sample=sample, s3_project=s3_project, panelname=panel)
+        cfpipeline(basespace_project_regex=basespace_project, sample_regex=sample, s3_project=s3_project, panelname=panel)
